Move Component previews under /spec
Zeitwerk will try to autoload everything under /app, which fails because
although these files are nested under 'previews' they do not define
parts of a Previews module.

Also remove a leading slash from the component preview path, which
causes the component previews to appear under rails/components again.
