+++ spec/components/previews/provider_interface/offer_summary_list_component_preview.rb
