+++ spec/components/previews/provider_interface/status_box_component_preview.rb
