+++ spec/components/previews/provider_interface/work_history_component_preview.rb
